Implement basic taskflow worker
A taskflow worker process, that runs flows from a persistent job board
in a parallel thread worker pool.  Tasks are defined through stevedore
plugins in the 'craton.workflow' namespace - see the TestFlow example.

Defaults are set to use an in-memory fake "persistent storage", for ease
of testing.  We will want to change this default at some point in the
future.

Also, the worker main() is currently hardcoded to feed in a single
TestFlow job.  Again, we will want to add a command-line tool and/or web
interface to post jobs once we have an out-of-process job board.