Refactor variables to use polymorphic association
Refactors VariableMixin such that it uses polymorphic association - a
'variable_association' table, that polymorphically links to resources
that are mixed in with variables; along with the actual variables
themselves in a 'variables' table. This replaces the need to use
'region_variables', 'cell_variables', 'label_variables', and
'device_variables' tables, and similar tables we would need to mix in
the future. In particular, User now has variables mixed in. Other
entities will use this functionality in the future.

For the new 'variables' table, 'key_' and 'value_' replace 'key' and
'value' columns. This renaming is only visible when using raw SQL -
otherwise the SQLAlchemy model maps this to the original names.

This change set is a refactoring - all existing code should work
without any change.

Change-Id: Ib34885fc132289c5fe892f17bbeecb42989f6810
Closes-Bug: 1614811