Adds search resources by variables
Each of regions/cells/hosts can be searched with
variables after this patch. So far any resource could
only be searched by defined parameters such as id
name, region_id etc. With this patch any query parameter
that is not a defined paratemeter will be passed in as
a variable parameter search.

Example:
/v1/hosts?key1=value1 will search hosts which has variables
key1 with value value1.

Closes Bug: 1611425

Change-Id: Ica0ac3a3ffde17dbc0cdb4e6bc55ddda27cc8e18