Refactor REST API support to remove boilerplate
This change removes the following boilerplate:

* Most filter setup, along with output to a JSON response, is now
  captured in the new craton.api.v1.base.filtered_context decorator.

* Provides standard exception to HTTP status code mapping using the
  new craton.api.v1.base.http_codes decorator.

There are additional opportunities for boilerplate removal, especially
in scenarios where we have multiple ways of looking up a specific cell
or region; but this boilerplate is not addressed in this fix.

This change does not implement related opportunities for cleanup, such
as import reordering or doc strings. These should be in separate
changes, per discussion on Gerrit.

In addition, this change now returns 204 upon DELETE in the API; and
adds version info to requirements.txt for Flask to conform to Jenkins
gate check.

Change-Id: I7da8e9620c2fb07ebd3953cc6d76d47c08651fe7
Partial-Bug: 1623095
Closes-Bug: 1628050