Remove code related to Python versions before 3.5
This commit removes any occurrence of the library six.

Six is used to aid in writing code that is compatible with both Python 2
and Python 3. Craton no longer supports the use of Python 2.x, and so
six is no longer required.

The minimum version of Python supported is 3.5, the classifiers are
updated accordingly in setup.cfg.

The docs have been updated to specify 3.5 as the minimum version.

Tox no longer attempts to test against Python 3.4.

This change also removes .travis.yml because it is left over from
from before the project was moved to the OpenStack project namespace and
is no longer required.

Closes-bug: #1630003
Change-Id: I63cba2b6f21a507c0ff02edbef24b663af400c12