Add tests to validate schema
This commit adds tests to validate the schema used by the API and
addresses any errors exposed by those tests.

Parts of the scheme currently do not conform to the required
specification, this leads to unexpected errors when attempting to use
parts of the API.

This commit addresses a bug in the definition of labels for
DefinitionsHost and DefinitionsHostId.

This commit addresses a bug in the definition of roles for
DefinitionUser.

The module jsonschema provides the ability to validate that a schema is
correctly constructed. The API should be able to assume that the schema
are valid. This commit adds tests that use jsonschema to validate all
the schema definitions.

Change-Id: Ia82bdfb8bf315681177f7d03ab4387c19ef86097