Refactor var resolution, blame to use common code
Computes resolution ordering, and corresponding blame, using a common
code base. This change simplifies testing, and makes it easy to
introduce resolution strategies for other entities besides devices.

Removes craton.db.api.device_blame_variables in favor of a blame
method on models that mixin VariableMixin.

A future fix will add project in the resolution ordering, but this
change was made as a refactoring, other than fixing blame support with
respet to labels.

Change-Id: I19a9f7569322c91ed4619fa8c923cb823349c196
Closes-Bug: #1637493