Remove filtered_context validation of requests
This change modifies filtered_context so that it no longer validates
that the request includes the required information.

filtered_context was duplicating the validation performed by
request_validate. This led to differences between what the schema and
filtered_context believed were required.

The schema is updated where it was missing required values that were
being supplied to filtered_context. Removing the reserved_keys means
there is no longer a difference between the key filtered_context uses to
identify an IP address and the one defined by the schema for GET and the
endpoints /hosts, /netdevices and /net_interfaces.

In addition, the 400 error in filtered_context was only raising a 400
error when 'not query_required and query_required is not None' or when
there was a difference between the required keys specified. In principle
this logic could be moved to the schema however, I'm not sure there is
any real benefit creating a special case for a region_id or device_id of
0 when there is already a code path for invalid IDs.

Change-Id: Ib459b9824815d6cac0eaa9e544e13b90fb61ecc1