Use Python 3.5 with tox
Update tox.ini so that pep8 testing uses a Python interpreter with
version 3.5. Without this fix the checks fail when using syntax that is
not supported by 2.7 because the gate defaults to 2.7.

basepython has been set in the parent section but can be overridden for
specific environments if required at a later date. This means all
enviroments now require 3.5 and not simply 3.

Change-Id: I22151eb1ac7398579ac9fb53c5ff1e664407ec93