Add device to net_interfaces endpoint
The net_interfaces endpoint is modified from:

/net_interfaces

to:

/devices/<device_id>/net_interfaces

This change only impacts the NetInterfaces resource.

This change means that it is no longer necessary to specify device_id as
part of the query string on a GET or in the data of a POST. This change
makes using the API more logical by creating a hierarchy in the
endpoint and simplifies enforcing requirements.

This change removes the need for filtered_context.

The result of these changes is that:

- the route variable rule limits device_id to int
- device_id is now supplied as an argument to the methods
- the GET query and POST request data is accessed through 'g'
- the context is in 'g'
- the schema has been modified so that device_id is no longer required
in requests but is still returned as part of the response
- the schema has been updated so that only properties with a genuine
default have one set, anything with a default set to None has had it
removed

The tests and generate_fake_data have been modified where necessary to
support the changes made.

This is one in a series of patches that will modify all endpoints to use
the same pattern.

Change-Id: I8f2d27ec0ff04779f909c259e96b7f34e82f895c
Partial-Bug: 1644503