Use local repo code when running functional tests
This change sets the path used by Docker so that code from the local
repo is used to build the docker image instead of
github.com/openstack/craton. This allows functional tests to be run
against local changes which mean functional tests can now be written at
the same time as the functionality changes.

Change-Id: I5c25910c1d96799da1f6afcaf0771568770c55fe