Remove required parameter from GET queries
This change modifies the API GET requests for listing resources so that
they no longer have query strings that include required attributes.

filtered_context and request_validate were both performing validation.
filtered_context no longer takes a 'required' argument and so will no
longer fail if the query string did not specify a particular attribute.
The schemata have been updated so that the GET requests no longer define
required properties.

In addition, the 400 error in filtered_context was only raising a 400
error when 'not query_required and query_required is not None' or when
there was a difference between the required keys specified. In principle
this logic could be moved to the schema however, I'm not sure there is
any real benefit creating a special case for a region_id or device_id of
0 when there is already a code path for invalid IDs.

Now that region_id is not a required parameter hosts_get_by_region does
not make sense, this commit replaces it with hosts_get_all which takes
region_id as an optional filter.

Removing the requirement for region_id when listing cells required the
DB API to be updated. This commit simplifies the request so that it
treats region_id as an additional filter and no longer attempts to
interpret the request based on the type of filters supplied.
cells_get_by_name has been removed, all requests to list cells now use
cells_get_all. To explicitly request a cell will a specific ID
/cells/<id> should be used. This is also in line with the other
inventory GET requests.

networks_get_by_region is replaced with networks_get_all, this makes the
function better reflect its purpose with the removal of region_id as a
required argument.

network_devices_get_by_region is replaced with network_devices_get_all
to better reflect the fact that a region_id is no longer required.

Similar to above, the removal of the requirement for a device_id meant
replacing network_interfaces_get_by_device with
network_interfaces_get_all. This creates a function that better
represents its purpose.

Unit tests have been modified and added to account for these changes.

The unit tests test_networks_get_by_name_filter_no_exit and
test_network_devices_get_by_filter_no_exit have been updated so that
they now perform a valid test, originally they were both querying an
empty database.

Change-Id: I0d69d16d59ed24ac4389137e89a462baf8e8269b