Fix filtering discrepancies
When providing filters for a GET request, the filter must exist as a
reserved_key in filtered_context, be part of the relevant schema, be
exposed in the DB query and be part of the model. This commit addresses
some differences between these different parts of the code base so that
the appropriate filters are available.

Change-Id: I076d677bac55475a509256206bd776f1a8cabfbd