Craton workflow: Modeling and API
- This patch is initial work for craton workflow
engine. Currently the process is like so:
1. Users can create task definition
2. Users can create their own workflow def from available
task def.
3. Tasks are RBAC based so certain tasks can be exposed
only to a certain group.
4. Generic workflows can also be made and exposed to users
by admin.

User submits a "job" of workflow def though the api. The job is
stored in the db, and is posted to a jobboard, workers pick
up the jobs from the jobboard and process it.

User can ask for what jobs he/she has running or ask for
status of job.

This patch only introduces the modeling and api for the
task/workflow and job creation.

Change-Id: I69f4b0de3112fb34e42a7d1e4e7a0d9d907e117f
Implements: blueprint workflow (partial)