Remove reserved_keys from filtered_context
reserved_keys is used to removed filters with a value of None, the
assumption being that those filters were not set in the query and have
added with the default value of None when the request was validated.

This commit removes the default values set for each GET request schema,
thereby preventing the validation process from adding them in if not
part of the query. This means there is no need for filtered_context to
filter them out because args only includes filters that were in the
query or those that have a useful default value.

Change-Id: I133fe349c4acc18f4678f4e6dd5e7d7558b21144