Return varialbes in response on create calls
If a user creates a region/cell/host and also passes variables
during the create process the variables get created
but are not present in the response.

This patch fixes that by adding variables in response
for a create calls on region/cell/hosts when variable
was in the create request body.

Closes Bug: #1648437

Change-Id: I19f78a07e7cd0bddcd3020f7b066a206a84af59a