Use Requests' Sessions in our Functional Test Suite
Using the Requests "functional" API (requests.{get,post,put,delete,..})
is inefficient as it causes a Session object, and several transport
adapters to be created for each request. Creating a session should
provide some performance benefit because we'll be creating fewer objects
and fewer sockets thanks to connection pooling.

This also corrects some of the assumptions in the test methods:

1. GET requests *can* have request bodies, but those are semantically
   undefined and may be stripped by intermediaries. So while services
   like Elasticsearch sometimes abuse this, it's not something we
   plan to do and as such we shouldn't use it in our tests.

2. We almost certainly don't want to set a Content-Type for every
   request. Even if we did, Requests would strip it on requests like a
   GET or DELETE that don't have a body.

   Further, when using the `json` keyword argument in a requests call,
   the library will generate the appropriate headers for us.

3. We can accept data in our delete call, but DELETEs are another
   request that don't assign semantic meaning to bodies.

Finally, we also allow each request to override headers as necessary and
set all of the default headers on the session object.

Change-Id: I284b0548e982cc4ca42323d47d604b5792d3ef1a