Add specification detailing Craton's URL design
After significant discussion, the Craton team has come to the conclusion
that we want to use a simple, flat, and concise URL structure for
current and future development. This specification provides the
reasoning for that decision and an enumeration of the work items
required for its successful implementation.

Describes-Blueprint: url-structure-and-design
Change-Id: I6f39bcf904e625bc39b0fa2936ac00864af8d90c