Refractor functional test for test isolation
The current functional test base sets up container
in a way that single container is used for all tests.
This means tests isolation is possible if we destroy the
db and recreate it for every test, instead this patch
changes the approach and makes each test spin up its own
continer, the teardown then removes the container.

Image is build for the module, so it does not need to be
setup on class or test level.

Cons of this approach is, because we spin up a new container
for each test, it slows the overall test run time.

Change-Id: Ia9020ccc0d012e5f2849b176911039525b1fb9f8