Disable Flask static directory
By default Flask assumes a directory called 'static' exists in the
application directory. Given that it doesn't exist in craton, attempting
to access files in this directory returns a 404 and HTML specifying the
error. Attempting to access other non-existent paths returns JSON
content.

The static directory is disabled because it is not required.

Change-Id: I13a339f433093115a0c364115f166174d6179204