Fix device variable data update validation
This commit addresses a number of issues with the way that variable data
updates were being validated:

- API PUT methods for setting device variables were not using the schema
validated data (g.json), they were using the raw data from the request
(request.json).

- DefinitionsData, the schema used to validate the variables data from
the request, was not properly validating the data, it expected the data
to be an object containing the names "key" and "value" with values of
type string and object.

- DefinitionVariablesSource, the schema used to filter the response
data, was not specifing all the appropriate types, null, boolean and
number were missing.

- The result of the above bugs is that in most cases, the API functioned
as expected however the code itself was misleading and any errors were
likely to cause confusion.

This commit modifies the PUT methods to use and ultimately return data
validated by the schema DefinitionVariablesSource.

Change-Id: I34c6b2e66bce6c52595a96ed4455429685f6220d