Use validated data in PUT requests
This commit updates a number of the put methods to ensure the data
supplied to any dbapi call in the method was first validated.

Tests have been added to ensure that invalid data is filtered out before
the dbapi call is made.

In addition, a number of the Flask request locations were incorrect in
the schema, this commit updates them so that the put methods are able to
correctly reference the data and tests have been added to prevent
regression.

Change-Id: Ib1cb016b539fe448d04930069c9cb2027469b7c9