Remove use of request.json in API methods
request.json is the raw JSON data supplied as part of an API request. If
this is used, it means the validation step is being bypassed. This
commit updates the API so that any occurrence of request.json is
replace with g.json, where g.json is the validated data.

Additional schemata have been added so that the request data can be
validated.

Change-Id: Icae6d5a2bcecee2b185b96c376033d18a882956a