Add context to function signature of all requests
This commit standardises the way that the context is accessed by methods
in the API classes. This change takes the approach in filtered_context
and applies it across all request methods so that they receive it as an
argument. This approach should simplify the functions, reduce the
duplication of code defining where the context is stored as well as
remove the need to access a global variable.

Change-Id: Id5dfe7b3ab5f71af915634a5b9b58272c120701d
Related-Bug: 1623095