Pass request data as an argument to API methods
This commit generalises the approach taken by filtered_context; the
methods no longer access the request data directly but instead receive
it as an argument.

This change helps to standardise the code base. Also, by automatically
supplying the data as an argument, it reduces the likelihood that the
validation process will be bypassed when any new functionality is added.
In earlier version, some parts of the API were directly accessing the
raw request data via 'request.json' instead of the validated data that
was in 'g.json'.

Change-Id: Icba9612ef5fa968de706ede610cce111ec9723c1
Related-Bug: 1623095