Log validated request data
This commit adds logging of validated request data. This change removes
the functionality from filtered_context and adds it to request_validate
so that any data supplied as part of a request body is also logged, in
addition to to logging of request parameters that was being done by
filtered_context.

Change-Id: Id830f7369463193b06c83c8b9e8533136a913cf5