Functional tests: add db teardown for isolation
The current functional test base sets up container
with all required services like api and db for all tests.
This adds to that by including a db teardown after each
test such that each test gets a new db but with project
and user intact, so each test can be isolated to its own.

The patch also removes generating data at start. Each test
has to create the resources it needs.

It also edits the http method signature for the tests. Since
headers are allowed to be overwritten, we need a way to not
have to pass it every call.

Change-Id: Ia9020ccc0d012e5f2849b176911039525b1fb9f8