Apply API-WG guidelines for naming resources
The URLs defined in the routes have been updated to replace underscores
with hyphens. The tests have been updated to account for this change. In
addition tests have been added to ensure that the URLs meet the
API-WG guidelines [1], namely that they contain only lowercase characters
and do not contain punctuation, with the exception of hyphens.

[1] https://specs.openstack.org/openstack/api-wg/guidelines/naming.html

Change-Id: I77c02de2973e12c6fbcdcc925fa6686aa1f7a094
Closes-Bug: 1657889