Require filter schema for every (endpoint, method)
response_filter is modified so that a request will fail if a filter is
not defined for its (endpoint, method). This complements the testing in
test_api_schema.py that ensures this has been done. These changes should
help to prevent filters being missed when the API is expanded.

The variable 'filter' has been renamed to 'resp_filter' so it does not
conflict with the built-in function filter.

No longer modify the status if a filter only supports one status code.
No filter currently is defined like this and this seems likely to cause
an error if one ever was.

If the filter does not define a schema for a particular status code,
this information should not be given to the client. This error is now
logged and a more general failure message is reported in the response.

Change-Id: Iee0435a33bf9ff96c05d33f5eefaf7adde8f073b
Partial-Bug: 1644270