Return variables object in response to PUT request
schemas.py is updated so that PUT requests to any of the variables
endpoints result in a response that includes an updated version of the
variables resource. Without this change these requests return the JSON
object null.

The tests have been updated, this included adjusting the DB mock so that
it returns an appropriately updated object. The mocks were returning
data that was not updated.

Change-Id: I267a59eca23343d5d8cd6b304dcf1102ada4225f