Fail if request includes unexpected parameter
This change modifies the behaviour of requests so that they now fail
with a 400 error if the request includes a parameter that is not defined
in its schema. Previously the behaviour was to ignore anything not
expected. This should help to prevent users being mislead by responses,
for example a GET request with filters that were not applied because the
query was incorrect.

Changing this behaviour also addresses a difference between the project
and the API-WG guidelines[1].

The JSON schema validation keyword additionalProperties is used to
ensure that undefined properties cause the request to fail.

GET requests that don't accept any form of query now have a schema so
that the validation can generate a failure if a query is specified.

[1] https://specs.openstack.org/openstack/api-wg/guidelines/http.html

Change-Id: I61b0d9445334f93d787a9330f11fa01e4e3503d4