Add generic variables API code and use for hosts
The API code for managing resource variables is essentially the same for
all types of resource. This commit seeks to reduce the duplication in
the hopes that it will bring benefits such as simplifying the
maintenance of the code base and reducing the potential for differences
between the different variables endpoints.

This change does not seek to modify the user experience of the API,
changes are limited to those required to create a standardised
implementation that can be consumed by all resources.

The change can be viewed as consisting of two broad steps:
- a generic variables API resource is added
- the hosts-specific variables implementation is removed and replaced by
  the generic implementation.

Separate patches will be submitted to convert the remaining variables
endpoints.

Change-Id: Id369dc32d9be33828cc34c9cfc67c36e5e7208dd
Partial-bug: 1661577