Fixes device search by label
- After commit aad5fd33c3577a67ebfdaadc3690b678ba364942
search for host devices by label was broken because the
pagination function was unable to handle joins on relationsip.

This patch switch to filter rule on query so that it palys
nice with the pagination patch.

Unit test on db and functional tests are added to test
filtering by label.

Closes Bug: 1662842

Change-Id: I476a5efcc1e5ef8735833dd37583653f2c9d56c5