Specify keys of variables to delete in REST API with array
DELETE on /v1/<resource>/<id>/variables currently takes a request in
which the *values* of the JSON object (dict) are the desired
keys. Although workable, this is at best counterintuitive, and
required manual inspection of the code to determine what the correct
format was.

Example of old usage without this fix:

curl http://127.0.0.1:8080/v1/hosts/8/variables -X DELETE \
  -H "Content-Type: application/json" -H "X-Auth-Token: demo" \
    -H "X-Auth-User: demo" -H "X-Auth-Project: b9f10eca66ac4c279c139d01e65f96b4" \
      -d '{"not-the-key-being-deleted": "x"}'

Instead, by using this proposed bug fix, this request will now return
this error message to the client, which is also informative:

{"errors": ["{'not-the-key-being-deleted': 'x'} is not of type 'array'"], "message": "Bad Request"}

Now with this change, the DELETE method now takes an array (list),
like so:

curl http://127.0.0.1:8080/v1/hosts/8/variables -X DELETE \
  -H "Content-Type: application/json" -H "X-Auth-Token: demo" \
    -H "X-Auth-User: demo" -H "X-Auth-Project: b9f10eca66ac4c279c139d01e65f96b4" \
      -d '["x", "y"]'

Such an API change better reflects that we are interested in deleting
a set or list of keys.

I don't believe the intent of the DELETE API was ever to be used in
this way, although this functionality was faithfully reproduced in the
recent refactoring of variables.

Closes Bug: 1663634

Change-Id: If63a88bc9df83a497ee99e50afdb1ad7aa3126a8