Add up-link to host responses
This commit updates the response object for hosts to include a links
property. Specifically the response object now includes a link to the
host's immediate ancestor, where that is either another device, a cell
or a region.

Example link:

{
    "rel": "up",
    "href": "http://localhost/v1/network-devices/1",
}

The link relation "up" is intended to represent "... a parent document
in a hierarchy of documents" [1] and so seems the most appropriate
choice of relation given resources are naturally ordered:
regions < cells < device < ... < device.

[1] http://www.iana.org/assignments/link-relations/link-relations.xhtml

Change-Id: I9f8e39984bad15eea72c7bea5af60907b95b65fc
Partial-bug: 1662496