Add up-link to network-device responses
This commit builds on the existing implementation of up-links for hosts.
It updates the response object for network-devices to include a links
property. Specifically the response object now includes a link to the
network-device's immediate ancestor, where that is either another
device, a cell or a region.

Example link:

{
    "rel": "up",
    "href": "http://localhost/v1/network-devices/1",
}

The link relation "up" is intended to represent "... a parent document
in a hierarchy of documents" [1] and so seems the most appropriate
choice of relation given resources are naturally ordered:
regions < cells < device < ... < device.

[1] http://www.iana.org/assignments/link-relations/link-relations.xhtml

Change-Id: I6b2a215db5ef1b3275fcbd0cf2ab5bd4d3996f03
Partial-bug: 1662496