WIP: Variable search for resources now uses resolved variables.
Implements a filter addition for vars such that each key-value pair in
the vars is searched for a possible match, regardless of associated
resource; then chases the associated resources for the specific
resource being searched for, thereby implementing the resolution
algorithm (but in reverse). Finally all such resources are checked to
see if they intersect with respect to their rooting key-value pairs in
the original search.

Examples:

$ craton-get v1/devices?vars=openstack_release:juno

$ craton-get v1/devices?vars=foo:abc,bar:xyz

A this time, resolution is always done; instead, it should be possible
to turn off resolution in the search query. (Should be fixed before
merging this patch in.)

Filter values are not currently encoded for JSON; but this will fixed
in a subsequent patch that is addressing generalized JSON matching via
JSON path.

Change-Id: I1d40d734e60b5563dfb01da05ffb6494ed9a919c
Closes-bug: 1661226
Closes-bug: 1669493