WIP: Variable search for resources now uses resolved variables.
FIXME some brief comments on how this algorithm is implemented

Examples:

$ craton-get v1/devices?vars=openstack_release:juno

$ craton-get v1/devices?vars=foo:abc,bar:xyz

A this time, resolution is always done; instead, it should be possible
to turn off resolution in the search query. (Fix before merging this
patch.)

Filter values are not currently encoded for JSON, so only literal
strings can be matched, not arbitrary JSON values. This should be
fixed, or at least partially relaxed to numbers and booleans, before
merging this patch in.

Change-Id: I1d40d734e60b5563dfb01da05ffb6494ed9a919c
Closes-bug: 1661226
Closes-bug: 1669493