Validate parent_id is neither id nor a descendant
A device's parent_id represents the id of another device. A parent_id
cannot be a device's own id nor can it be the id of a descendant.

This commit adds a validator decorator to ensure that the parent_id
supplied meets the requirements before the database is updated. Failure
to properly validate the parent_id can cause the database to become
unresponsive.

If an invalid parent_id is supplied the API will return a 400 Bad
Request.

Change-Id: I81f6bff5bf64b7c441fef08a8e03731cc17bb2f1
Closes-bug: 1666536