Create user under project_id passed in request
Currently when user is created, its taking the project_id
parameter from the user who is creating the request.
However, given root users are expected to be ablet to create
users for all projects this patch modifies the behaviour
and ensure that the user creation is always under the
project_id parameter that is passed in.

Closes Bug: 1666695

Change-Id: I19fb901ad41edeeb0a92f41264867a768a207fce