Return variables in response on Network Device create
If a user creates a network device and also passes
variables during the create process the variables
get created but are not present in the response.

This patch fixes that by adding variables in response
for a create calls on network devices when variable
was in the create request body.

Change-Id: Id9818d5c30a9a0c17039191335f1b75525003a3d
Closes-Bug: 1675087